requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
jinja2
Pillow